fix for heartbeat sending multiple requests at once due to dual loading the module it was a part of
