+++ components/pages/chart-preview.tsx
@@ -18,6 +18,9 @@ import { ModeToggle } from "@/components/dark-mode-toggle";
 import { calculateSignalPercentage } from "@/utils/signalMetrics";
 import { ArrowRightIcon } from "@radix-ui/react-icons";
 
+import { useAuth } from "@/hooks/auth";
+import heartbeat from "@/hooks/heartbeat";
+
 interface ModemResponse {
   response: string;
 }
@@ -57,8 +60,12 @@ export default function ChartPreviewSignal() {
   });
   const [initialLoading, setInitialLoading] = useState(true);
   const previousData = useRef<SignalData | null>(null);
-
+  const { logout } = useAuth();
+  const { isServerAlive } = heartbeat();
   useEffect(() => {
+    if (!isServerAlive) {
+      logout();
+    }
     const fetchStats = async () => {
       try {
         const response = await fetch("/cgi-bin/quecmanager/at_cmd/fetch_data.sh?set=5");
@@ -125,7 +132,7 @@ export default function ChartPreviewSignal() {
     fetchStats();
     const intervalId = setInterval(fetchStats, 2000);
     return () => clearInterval(intervalId);
-  }, [initialLoading]);
+  }, [initialLoading, isServerAlive, logout]);
 
   const chartData: ChartDataItem[] = [
     {
